Add files via upload
The purpose of the project is to apply the concepts learned in the Agile Scrum Master course by analyzing a hypothetical organization.GTM systems is a large IT company with offices all around the world. The company delivers software products and services to corporate clients. While GTM has no expertise in meteorology, it proposes to aggregate weather data from multiple providers and uses analytics to correlate it with meaningful conclusions for businesses. the project encompasses Backlog management, epics, user stories and reports created in JIRA.